+++ crates/wdl/src/lib.rs
@@ -97,6 +97,9 @@ pub use wdl_lint as lint;
 #[doc(inline)]
 pub use wdl_lsp as lsp;
 
+#[cfg(feature = "cli")]
+pub mod cli;
+
 #[cfg(test)]
 mod test {
     use std::collections::HashSet;